Try bumping JuliaSyntax to 0.4.10
The last time we tried to bump JuliaSyntax (https://github.com/JuliaLang/julia/pull/53119),
it didn't work out due to test failues. As a result, we haven't actually
bumped JuliaSyntax in about a year and there are several syntax changes
in JuliaSyntax that we haven't picked up yet (i.e. are waiting in PRs).
I don't know what the current situation is, but I think we need to get
back to a point where we can bump JuliaSyntax regualarly, so let's try
to take the first step.