compute Matrix length from dims
Since changing Array to use Memory as the backing, we had the option of
making non-Vector arrays more flexible, but had instead preserved the
restriction that they must be zero offset and equal in length to the
Memory. This results in extra complexity, restrictions, and allocations
however, but doesn't gain any known benefits. This PR aims to test if
nanosoldier detects any benefit, or whether this restriction has
outlived its usefulness.