inference: fix backedge correctness for `applicable` calls
Even if the return value of `applicable(f, args...)` is initially
inferred as `Const(true)`, when a new method is added to `f` then it may
cause new method match ambiguities, requiring the call to return return
`false` at the next invocation. To handle such cases, it is necessary to
always add a method table backedge in the inference of `applicable`.

Also just as a minor backedge reduction optimization, this commit avoids
adding backedges when `applicable` is inferred to return `::Bool`.